Conflicts:
	WebContent/WEB-INF/web.xml
	WebContent/guestbook.jsp
	WebContent/index.jsp
	WebContent/manage/guestbook.jsp
	src/cn/yh/easybuy/web/CommentServlet.java
	src/db.properties

+++ src/cn/yh/easybuy/entity/Product.java
@@ -5,7 +5,7 @@ public class Product {
 	private Integer id;
 	private String name;
 	private String description;
-	private Float price;
+	private Double price;
 	private Integer stock;
 	private Integer cid;
 	private Integer childId;
@@ -28,10 +28,11 @@ public class Product {
 	public void setDescription(String description) {
 		this.description = description;
 	}
-	public Float getPrice() {
+
+	public Double getPrice() {
 		return price;
 	}
-	public void setPrice(Float price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 	public Integer getStock() {
